frontend/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
httpx==0.28.1
psycopg2-binary==2.9.10
greenlet==3.1.1
google-genai==1.68.0
